feat: add a basic RFC7468 codec
This gives us a standardized format to read and write transportable
files, even for our silly little internal formats.

Yes, this is a bit gross, and yes this is definitely error-prone, but
again, this is not production-grade and you shouldn't be using it for
anything serious. ☺

Includes a minimal roundtrip test to ensure function.

Signed-off-by: Sam Stuewe <samuel.stuewe@gmail.com>